Moved invalid billing banner
