+++ ui/src/app/lab/pages/workspace/invalid-billing-banner.tsx
